Refactor utils module and add bump allocator
Move C-like string conversion and number parsing functions into a
dedicated cstr module. Add a simple bump allocator implementation
in malloc.rs for dynamic memory allocation. Update mod.rs to
re-export these modules for easier access.
